+++ app/src/main/java/ru/practicum/android/diploma/di/ViewModelModule.kt
@@ -5,7 +5,6 @@ import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.core.parameter.parametersOf
 import org.koin.dsl.module
 import ru.practicum.android.diploma.details.presentation.viewmodel.VacancyDetailsViewModel
-import ru.practicum.android.diploma.search.ui.SearchViewModel
 import ru.practicum.android.diploma.search.presentation.viewmodel.SearchViewModel
 
 val viewModelModule = module {
@@ -19,6 +18,9 @@ val viewModelModule = module {
     }
 
     viewModel<SearchViewModel> {
-        SearchViewModel(interactor = get(),getSuggestsUseCase = get())
+        SearchViewModel(
+            interactor = get(),
+            getSuggestsUseCase = get()
+        )
     }
 }
