--- app/src/main/java/ru/practicum/android/diploma/search/ui/SearchViewModel.kt
@@ -1,22 +1,0 @@
-package ru.practicum.android.diploma.search.ui
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import kotlinx.coroutines.launch
-import ru.practicum.android.diploma.search.domain.api.GetSuggestionsForSearchUseCase
-
-class SearchViewModel(private val getSuggestsUseCase: GetSuggestionsForSearchUseCase) :
-    ViewModel() {
-    private var suggestionsList = MutableLiveData<List<String>>(emptyList())
-    val suggestionsLivaData: LiveData<List<String>> = suggestionsList
-    fun getSuggestionsForSearch(textForSuggests: String) {
-        viewModelScope.launch {
-            getSuggestsUseCase.execute(textForSuggests)
-                .collect {
-                    suggestionsList.postValue(it)
-                }
-        }
-    }
-}
